requirements.txt
crewai
crewai-tools
litellm
python-dotenv